Moved feature importance estimation methods into one file
